Changed all doubles to floats. Debug: 118 to 98 fps Release: 20 to 15 fps

+++ GameParticles/Math/Vect4D.cpp
@@ -12,7 +12,7 @@ Vect4D::Vect4D()
 	this->w = 1.0;
 }
 
-Vect4D::Vect4D( double tx, double ty, double tz, double tw )
+Vect4D::Vect4D( float tx, float ty, float tz, float tw )
 {
 	this->x = tx;
 	this->y = ty;
@@ -27,7 +27,8 @@ Vect4D::~Vect4D()
 
 void Vect4D::norm(Vect4D& out)
 {
-	double mag = sqrt( this->x * this->x + this->y * this->y + this->z * this->z );
+	//THIS WILL NEED TO BE CHANGED TO AVOID SQUARE ROOT
+	float mag = sqrtf( this->x * this->x + this->y * this->y + this->z * this->z );
 
 	if( 0.0f < mag )
 	{
@@ -60,7 +61,7 @@ Vect4D Vect4D::operator - ( Vect4D t )
 	return out;
 }
 
-Vect4D Vect4D::operator *( double scale)
+Vect4D Vect4D::operator *( float scale)
 {
 	Vect4D tmp;
 
@@ -71,7 +72,7 @@ Vect4D Vect4D::operator *( double scale)
 	return tmp;
 }
 
-double& Vect4D::operator[]( VECT_ENUM e )
+float& Vect4D::operator[]( VECT_ENUM e )
 {
 	switch(e)
 	{
@@ -101,7 +102,7 @@ void Vect4D::Cross( Vect4D& vin, Vect4D& vout)
 	vout.w = 1.0f;
 }
 
-void Vect4D::set(double tx, double ty, double tz, double tw)
+void Vect4D::set(float tx, float ty, float tz, float tw)
 {
 	this->x = tx;
 	this->y = ty;
