+++ GameParticles/ParticleEmitter.h
@@ -27,12 +27,12 @@ private:
 	Vect4D	start_position;
 	Vect4D	start_velocity;
 	
-	double	max_life;
+	float	max_life;
 	int		max_particles;
 	int		last_active_particle;
-	double	spawn_frequency;
-	double	last_spawn;
-	double	last_loop;
+	float	spawn_frequency;
+	float	last_spawn;
+	float	last_loop;
 
 
 	// added for speed efficiency
@@ -40,7 +40,7 @@ private:
 	Particle *headParticle;
 	Vect4D	vel_variance;
 	Vect4D	pos_variance;
-	double	scale_variance;
+	float	scale_variance;
 	std::list<Particle> drawBuffer;
 	std::list<Particle> particle_list;
